fixed a small bug

+++ plugins/tickettool.py
@@ -17,7 +17,7 @@ class Tickettool(commands.Cog):
             return
 
         member_name = channel.name.replace("-", "#")
-        
+
         f = open(f"{channel.name}_{channel.guild.id}.html", "w", encoding="utf-8")
         f.write("<html>\r\n<link rel=\"stylesheet\" href=\"https://meyerweb.com/eric/tools/css/reset/reset200802.css\"><link rel=\"preconnect\" href=\"https://fonts.googleapis.com\"><link rel=\"preconnect\" href=\"https://fonts.gstatic.com\" crossorigin><link href=\"https://fonts.googleapis.com/css2?family=Open+Sans&display=swap\" rel=\"stylesheet\">\r\n<style>body{font-family: 'Open Sans', sans-serif;background-color:#36393f;color:#ffffff;width:98%;padding:15px 15px}</style>\r\n")
         f.write(f"<header style=\"display:flex;flex-direction:row;margin-bottom:20px\"><img src=\"{channel.guild.icon_url}\" width=\"60\" height=\"60\"><span style=\"font-size:34px;padding:15px 8px\">{channel.guild.name} - {member_name}</span></header>\r\n<body>")
@@ -233,8 +233,9 @@ class Tickettool(commands.Cog):
         message = await channel.fetch_message(payload.message_id)
         if str(payload.emoji) == "🔒" and payload.member != self.client.user and len(message.embeds) > 0:
             if message.embeds[0].title == "New ticket created!" and message.embeds[0].description == "Moderators can close the ticket by adding the :lock: reaction to this message.\nPlease don't change the name of the channel and his topic :warning:" and message.embeds[0].colour == discord.Color(MAINCOLOR):
-                await self.gen_html_close(channel, payload.member)
-                await channel.delete()
+                if message.channel.name != f"{payload.member.name}-{payload.member.discriminator}" and message.channel.topic != f"{payload.member.id}":
+                    await self.gen_html_close(channel, payload.member)
+                    await channel.delete()
 
 
 def setup(client):
